Refactor core module structure
