+++ include/candle_manager.h
