+++ src/signal.cpp
@@ -1,0 +1,1 @@
+// TODO: Implement signal processing logic
